refactor(knowledge): move overlay parser tests to a sibling file, silence expected dead-code
